pipeline/est/simulation_initial_conditions/requirements.txt
git+https://github.com/COVID-IWG/epimargin@master#egg=epimargin
appnope==0.1.0
arviz==0.9.0
astroid==2.4.1
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.9.3
certifi==2020.6.20
cftime==1.1.3
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
cligj==0.5.0
colorama==0.4.3
commonmark==0.9.1
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
descartes==1.1.0
fastprogress==0.2.3
flake8==3.8.3
Flask==1.1.2
flat-table==1.1.1
google-cloud-storage==1.7.0
h5py==2.10.0
idna==2.9
importlib-metadata==1.6.1
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.3
joblib==0.14.1
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
linearmodels==4.19
lxml==4.6.3
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
munch==2.5.0
mypy-extensions==0.4.3
netCDF4==1.5.3
numpy==1.18.2
packaging==20.4
pandas==1.0.3
parso==0.7.0
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.1
pprintpp==0.4.0
prompt-toolkit==3.0.5
property-cached==1.6.4
ptyprocess==0.6.0
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.7.4
pyhdfe==0.1.0
pylint==2.5.2
pymc3==3.9.3
pyparsing==2.4.7
pyproj==2.6.0
pyreadr==0.4.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
rich==2.3.0
rope==0.17.0
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.10.0
Shapely==1.7.0
six==1.14.0
sklearn==0.0
soupsieve==2.2.1
statsmodels==0.11.1
Theano==1.0.5
tikzplotlib==0.9.4
toml==0.10.1
tqdm==4.45.0
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.9
urlpath==1.1.7
wcwidth==0.1.9
Werkzeug==1.0.1
wrapt==1.12.1
xarray==0.15.1
xlrd==1.2.0
zipp==3.1.0